chore: move s3 functions from script to util
