+++ crawler/load_to_s3.py
@@ -1,6 +1,3 @@
-import boto3
-from dotenv import load_dotenv
-import os
 from datetime import datetime, timezone, timedelta
 
 from crawler.crawl_musinsa import (
@@ -9,35 +6,7 @@ from crawler.crawl_musinsa import (
     get_soup_object_from_html,
 )
 from crawler.list_crawler import get_goods_url_from_list_page
-
-load_dotenv()
-
-
-def get_s3_connection():
-    try:
-        s3 = boto3.client(
-            service_name="s3",
-            region_name=os.getenv("aws_region"),
-            aws_access_key_id=os.getenv("aws_access_key"),
-            aws_secret_access_key=os.getenv("aws_secret_key"),
-        )
-    except Exception as e:
-        print(e)
-    else:
-        print("s3 bucket connected!")
-        return s3
-
-
-def upload_html_to_s3(s3, html, s3_path):
-    try:
-        s3.put_object(
-            Bucket=os.getenv("s3_bucket_name"),
-            Key=s3_path,
-            Body=html,
-            ContentType="text/html",
-        )
-    except Exception as e:
-        print(e)
+from util.s3 import get_s3_connection, upload_html_to_s3
 
 
 if __name__ == "__main__":
